+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './object-utils';
 export * from './array-dimension-utils';
 export * from './array-random-utils';
 export * from './array-separate-utils';
+export * from './array-diff-utils';